requirements.txt
asgiref==3.4.1
backcall==0.2.0
Bottleneck==1.3.5
certifi==2022.06.15
decorator==5.1.1
Django==3.2.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
executing==0.8.3
jedi==0.18.1
joblib==1.1.0
numexpr==2.8.3
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.2
scipy==1.9.0
setuptools==61.2.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
stack-data==0.2.0
threadpoolctl==3.1.0
traitlets==5.1.1
typing_extensions==4.3.0
wcwidth==0.2.5
wheel==0.37.1
